Extract inline `util` mod to folder mod
